add (x,y) values to csv. omit unuseful comments.
I checked to see if the rgb values are accurate
per (x,y) location. They are. Going to a pixel
with gimp would reveal the exact same rgb values
as my csv file. Though I used int's and rgb values
seem to be stored as floats. Updating this would
be trivial. I will do so if I need to. Though I
image that would come with non-trivial computational
overhead.

It should be noted. Mapping (x,y) points from this
to a cartesian coordinate system has a twist.
The very top left pixel is (0,0). Which results in the
picture being placed in the 4th quadrant where we are
seeing the absolute value of y. In other word the point
(x,y) in a picture would translate to (x,|y|) on a coordinate
system.